+++ packages/sdk/electron/src/platform/ElectronPlatform.ts
@@ -5,7 +5,7 @@ import ElectronCrypto from './ElectronCrypto';
 import ElectronEncoding from './ElectronEncoding';
 import ElectronInfo from './ElectronInfo';
 import ElectronRequests from './ElectronRequests';
-import ElectronStorage from './ElectronStorage';
+import { getElectronStorage } from './ElectronStorage';
 
 // NOTE: Because Electron main process runs on Node.js, this platform should be
 // very similar to the Node server sdk platform.
@@ -21,9 +21,32 @@ export default class ElectronPlatform implements platform.Platform {
 
   requests: platform.Requests;
 
-  constructor(logger: LDLogger, clientSideId: string, options: ElectronOptions) {
-    const namespace = this.crypto.createHash('sha256').update(clientSideId).digest?.('base64url');
-    this.storage = new ElectronStorage(namespace!, logger);
+  constructor(logger: LDLogger, options: ElectronOptions) {
+    const globalStorage = getElectronStorage();
+    this.storage = {
+      async get(key: string): Promise<string | null> {
+        try {
+          return await globalStorage.get(key);
+        } catch (error) {
+          logger.error(`Error getting key from storage: ${key}, reason: ${error}`);
+          return null;
+        }
+      },
+      async set(key: string, value: string): Promise<void> {
+        try {
+          await globalStorage.set(key, value);
+        } catch (error) {
+          logger.error(`Error setting key in storage: ${key}, reason: ${error}`);
+        }
+      },
+      async clear(key: string): Promise<void> {
+        try {
+          await globalStorage.clear(key);
+        } catch (error) {
+          logger.error(`Error clearing key from storage: ${key}, reason: ${error}`);
+        }
+      },
+    };
     this.requests = new ElectronRequests(
       options.tlsParams,
       options.proxyOptions,