+++ packages/sdk/react/src/server/LDIsomorphicProvider.tsx
@@ -41,7 +41,7 @@ export interface LDIsomorphicProviderProps {
  * **NOTE:** This component is designed to be used in conjunction with {@link LDIsomorphicClientProvider}
  * in a server component to compute the bootstrap data and render this provider automatically.
  *
- * See the `server-only` example for how to use this component.
+ * See the `react-server-example` example for how to use this component.
  */
 export async function LDIsomorphicProvider({
   session,