+++ packages/shared/sdk-client/src/LDClientImpl.ts
@@ -45,6 +45,7 @@ import {
 } from './context/createActiveContextTracker';
 import { ensureKey } from './context/ensureKey';
 import { DataManager, DataManagerFactory } from './DataManager';
+import type { InternalDataSystemOptions } from './datasource/LDClientDataSystemOptions';
 import createDiagnosticsManager from './diagnostics/createDiagnosticsManager';
 import {
   createErrorEvaluationDetail,
@@ -83,6 +84,8 @@ export default class LDClientImpl implements LDClient, LDClientIdentifyResult {
   private _eventSendingEnabled: boolean = false;
   private _baseHeaders: LDHeaders;
   protected dataManager: DataManager;
+  protected readonly isFDv2: boolean;
+  protected readonly dataSystemConfig?: InternalDataSystemOptions;
   protected readonly environmentMetadata: LDPluginEnvironmentMetadata;
   private _hookRunner: HookRunner;
   private _inspectorManager: InspectorManager;
@@ -161,6 +164,8 @@ export default class LDClientImpl implements LDClient, LDClientIdentifyResult {
       this.emitter,
       this._diagnosticsManager,
     );
+    this.isFDv2 = !!this._config.dataSystem;
+    this.dataSystemConfig = this._config.dataSystem;
 
     const hooks: Hook[] = [...this._config.hooks];