+++ packages/shared/sdk-server/__tests__/data_sources/OneShotInitializerFDv2.test.ts
@@ -63,7 +63,6 @@ describe('given a one shot initializer', () => {
       initMetadata: undefined,
       payload: {
         type: 'full',
-        id: `mockId`,
         state: `mockState`,
         updates: [
           {