+++ packages/shared/sdk-server/__tests__/data_sources/StreamingProcessorFDv2.test.ts
@@ -248,7 +248,6 @@ describe('given a stream processor with mock event source', () => {
       },
       payload: {
         type: 'full',
-        id: `mockId`,
         state: `mockState`,
         updates: [
           {