+++ packages/sdk/server-node/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `@launchdarkly/node-server-sdk` will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [9.10.11](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.10.10...node-server-sdk-v9.10.11) (2026-04-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.18.3 to 2.18.4
+
 ## [9.10.10](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.10.9...node-server-sdk-v9.10.10) (2026-03-23)