+++ packages/store/node-server-sdk-redis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [4.2.23](https://github.com/launchdarkly/js-core/compare/node-server-sdk-redis-v4.2.22...node-server-sdk-redis-v4.2.23) (2026-04-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/node-server-sdk bumped from 9.10.10 to 9.10.11
+  * peerDependencies
+    * @launchdarkly/node-server-sdk bumped from >=9.4.3 to >=9.10.11
+
 ## [4.2.22](https://github.com/launchdarkly/js-core/compare/node-server-sdk-redis-v4.2.21...node-server-sdk-redis-v4.2.22) (2026-03-23)